==> notebooks/language/03.0-Buckeye-prep-dataset.ipynb <==
# Prepare Buckeye (English dataset)
1. grab linguistic units from dataset
2. grab dataset statistics

import numpy as np
import pandas as pd
from glob import glob
from parallelspaper.speech_datasets import prep_BUCKEYE
from parallelspaper.config.paths import DATA_DIR

BUCKEYE_DIR = '/mnt/cube/Datasets/buckeye/'

### Load data

speaker_list = glob(BUCKEYE_DIR + '*.zip')

(track_durations, word_durations, utterance_durations, phone_durations,
 all_words, all_phonemic, all_phonetic, all_pos) = prep_BUCKEYE(speaker_list)

### Get dataset statistics

def flatlist(list_of_lists):
    return [val for sublist in list_of_lists for val in sublist]

num_phonemes = len(phone_durations)
num_words = len(word_durations)
word_durations_s = word_durations
word_length_phones = [len(i) for i in flatlist(np.concatenate(all_phonemic))]
phone_duration_s = phone_durations
unique_phones = len(np.unique(flatlist(flatlist(flatlist(all_phonemic)))))
unique_words = len(np.unique(flatlist(flatlist(all_words))))
utterance_length_phones = [len(i) for i in flatlist(all_phonemic)]
n_sessions = len(all_phonemic)
session_durations = track_durations
total_duration = np.sum(track_durations)

stats_df = pd.DataFrame([[
        num_phonemes,
        num_words,
        word_durations_s,
        word_length_phones,
        phone_duration_s,
        unique_phones,
        unique_words,
        utterance_length_phones,
        n_sessions,
        session_durations,
        total_duration
    ]],
    columns = [
        'num_phonemes',
        'num_words',
        'word_durations_s',
        'word_length_phones',
        'phone_duration_s',
        'unique_phones',
        'unique_words',
        'utterance_length_phones',
        'n_sessions',
        'session_durations',
        'total_duration'
        ])
stats_df

# statistics for this language
stats_df.to_pickle((DATA_DIR / 'stats_df/BUCKEYE_stats_df.pickle'))

### make sequence dataframes

# all_words, all_phonemic, all_phonetic, all_pos

seq_df = pd.DataFrame(columns = ['language', 'levels', 'data'])

seq_df.loc[len(seq_df)] = ['english', 'speaker/utterance/word', all_words]
seq_df.loc[len(seq_df)] = ['english', 'speaker/utterance/word/phonemic', all_phonemic]
seq_df.loc[len(seq_df)] = ['english', 'speaker/utterance/word/phonetic', all_phonetic]
seq_df.loc[len(seq_df)] = ['english', 'speaker/utterance/pos', all_pos]

seq_df.to_pickle((DATA_DIR / 'speech_seq_df/BUCKEYE_seq_df.pickle'))
